Redesign the review view
- The heading now simply says "Capture Review" instead of "Review
  Capture List".
- The horizontal spacing between the heading and the skip button has
  been doubled.
- The progress indicator now talks of items instead of reviews, e.g.
  "Item 1 of 3" and has a custom font color
- There's now three options to choose from, labeled (from top to bottom)
  "Schedule", "Postpone" and "Delete". They all share the same width,
  but decrease in height to subtly encourage the user to schedule items
  over postponing them over deleting them.
- The two calendar widget have been merged into one that is updated
  depending on which option the user selected in that first step.
- There's a small close button in the top right of the widget to go back
  to the first step, if necessary. Pressing ESC also works!
- When picking a date there's now a brief delay of 0.5 seconds between
  hiding the calendar widget and showing the buttons again to give the
  user time to realize that the reviewed item has changed.
- Pressing "Delete" now requires a confirmation by pressing the button
  again. When pressing it for the first time, the button's text will
  change to "Are you sure?" and the button will turn red. Moving the
  mouse away from the button will reset those changes again.
- Since the top bar isn't visible while inside the review view, it now
  contains a duplicate of the dark/light mode toggle there.

+++ main_window/main_panel/review_view/review_view.gd
@@ -14,29 +14,28 @@ func _ready() -> void:
 
 func _setup_initial_state() -> void:
 	_on_main_panel_changed()
+	_on_dark_mode_changed()
 
 
 func _connect_signals() -> void:
 	#region Global Signals
 	Settings.main_panel_changed.connect(_on_main_panel_changed)
+
+	Settings.dark_mode_changed.connect(_on_dark_mode_changed)
 	#endregion
 
 	#region Local Signals
+	$Mode.pressed.connect(_on_mode_pressed)
+
 	%SkipButton.pressed.connect(switch_to_list_view)
 
-	%Schedule/CalendarWidget.day_button_pressed.connect(func(date):
-		_schedule_current_item(date.day_difference_to(DayTimer.today))
-		%Schedule/CalendarWidget.reset_view_to_today()
-	)
+	%Options/Schedule.pressed.connect(_show_schedule_calendar)
+	%Options/Postpone.pressed.connect(_show_postpone_calendar)
+	%Options/Delete.pressed.connect(_delete_current_item)
+	%Options/Delete.mouse_exited.connect(_reset_delete_button)
 
-	$DeleteButton.pressed.connect(_delete_current_item)
-
-	%Postpone/CalendarWidget.day_button_pressed.connect(func(date):
-		_postpone_current_item(date.day_difference_to(DayTimer.today))
-		%Postpone/CalendarWidget.reset_view_to_today()
-	)
+	%Calendar/FirstRow/CloseButton.pressed.connect(_hide_calendar)
 	#endregion
-
 
 func _on_main_panel_changed() -> void:
 	visible = (Settings.main_panel == Settings.MainPanelState.CAPTURE_REVIEW)
@@ -81,7 +80,7 @@ func _start_review() -> void:
 func _review_next_item() -> void:
 	current_review_id += 1
 	if current_review_id < review_queue.size():
-		%Progress.text = "Review %d of %d:" % [
+		%Progress.text = "Item %d of %d:" % [
 			current_review_id + 1,
 			total_reviews_due
 		]
@@ -90,32 +89,101 @@ func _review_next_item() -> void:
 		switch_to_list_view()
 
 
-func _schedule_current_item(day_offset: int) -> void:
+func _schedule_current_item(date: Date) -> void:
 	Cache.move_item(
 		review_queue[current_review_id].line_id - scheduled_items,
 		"capture",
-		DayTimer.today.add_days(day_offset).as_string(),
+		date.as_string()
 	)
 	scheduled_items += 1
+	_hide_calendar(true)
 	_review_next_item()
 
 
-func _postpone_current_item(day_offset: int) -> void:
+func _postpone_current_item(date: Date) -> void:
 	Cache.postpone_item(
 		review_queue[current_review_id].line_id,
 		"capture",
-		DayTimer.today.add_days(day_offset).as_string(),
+		date.as_string()
 	)
+	_hide_calendar(true)
 	_review_next_item()
 
 
 func _delete_current_item() -> void:
-	Cache.delete_item(
-		review_queue[current_review_id].line_id,
-		"capture"
-	)
-	_review_next_item()
+	if %Options/Delete.text == "Delete":
+		%Options/Delete.text = "Are you sure?"
+		var style_box := StyleBoxFlat.new()
+		style_box.bg_color = Color("#BF616A")
+		%Options/Delete.add_theme_stylebox_override("hover", style_box)
+	else:
+		Cache.delete_item(
+			review_queue[current_review_id].line_id,
+			"capture"
+		)
+		_reset_delete_button()
+		_review_next_item()
 
 
 func switch_to_list_view() -> void:
 	Settings.set_deferred("main_panel", Settings.MainPanelState.LIST_VIEW)
+
+
+func _show_schedule_calendar() -> void:
+	%Options.hide()
+	%Calendar.show()
+	%Calendar/FirstRow/Heading.text = "Schedule for…"
+	%Calendar/Widget.day_button_tooltip = "Click to delay decision until this date"
+	%Calendar/Widget.include_today = true
+
+	if %Calendar/Widget.day_button_pressed.is_connected(_postpone_current_item):
+		%Calendar/Widget.day_button_pressed.disconnect(_postpone_current_item)
+	if not %Calendar/Widget.day_button_pressed.is_connected(_schedule_current_item):
+		%Calendar/Widget.day_button_pressed.connect(
+			_schedule_current_item
+		)
+
+	%Calendar/Widget.reset_view_to_today()
+
+
+func _show_postpone_calendar() -> void:
+	%Options.hide()
+	%Calendar.show()
+	%Calendar/FirstRow/Heading.text = "Ask again…"
+	%Calendar/Widget.day_button_tooltip = "Click to move to-do to this date"
+	%Calendar/Widget.include_today = false
+
+
+	if %Calendar/Widget.day_button_pressed.is_connected(_schedule_current_item):
+		%Calendar/Widget.day_button_pressed.disconnect(_schedule_current_item)
+	if not %Calendar/Widget.day_button_pressed.is_connected(_postpone_current_item):
+		%Calendar/Widget.day_button_pressed.connect(
+			_postpone_current_item
+		)
+
+	%Calendar/Widget.reset_view_to_today()
+
+
+func _hide_calendar(pause := false) -> void:
+	%Calendar.hide()
+	if pause:
+		await get_tree().create_timer(0.5).timeout
+	%Options.show()
+
+
+func _reset_delete_button() -> void:
+	%Options/Delete.text = "Delete"
+	%Options/Delete.remove_theme_stylebox_override("hover")
+
+
+func _on_mode_pressed() -> void:
+	Settings.dark_mode = not Settings.dark_mode
+
+
+func _on_dark_mode_changed() -> void:
+	if Settings.dark_mode:
+		$Mode.icon = preload("images/mode_light.svg")
+		$Mode/Tooltip.text = "Switch To Light Mode"
+	else:
+		$Mode.icon = preload("images/mode_dark.svg")
+		$Mode/Tooltip.text = "Switch To Dark Mode"
